pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "titan"
version = "0.1.0"
description = "Titan paper-first to live trading system scaffold"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "pandas>=2.2",
  "numpy>=1.26",
  "rich>=13.7",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
titan = "titan.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
